+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## [1.3.1] - ??? v. 1.3.1
+### Deprecated
+- `function cpv_do ( $var )` better use `vars_do`
+- `[cpv]` better use `[vars]`
+### Fixes
+- Ugly row meta when GitHub Updater is active
+- Fixed bug with MCE button with html code in var
+
 ## [1.3.0] - 2019-10-29 v. 1.3.0
 ### Plugin renamed from cpvars to vars
 When you upgrade from a previous version to 1.3.0, please deactivate and reactivate the plugin to be sure that the migration is complete.